backend/requirements.txt
Django==3.2
djoser==2.1.0
djangorestframework==3.13.1